Fixes: Remove faulty constness for QMap in DuiInputContext
RevBy: Pekka Vuorela
Details: The IC tests can crash because of faulty constness that was used when
filling the QMap, probably resulting in the wrong (that is, const) operator[]
overload to be selected. It turned out that most temporarily created variables
in DuiInputContext::getStateInformation() can be easily avoided, which was done
with this commit.

Full, original commit at 57cc0b2ada90fbb97f96d5162a196e89896e3b44 in maliit-framework