Keep Maliit and MInputMethod namespaces separate.
RevBy: Jan Arne Petersen
Details: Gets rid of complicated the logic that tried to
avoid duplicating the code. A define was used in the public header
as a conditional for this logic, so it would only work
if the consumers of the headers had this define correctly when
building.

Full, original commit at 353b8a51fd6c9b770ae95e9763aa5afc8f4b72d8 in maliit-framework