Cache "x" argument to mouseHandler() as "preeditClickPos" in widget info.
RevBy: Kimmo Surakka, Pekka Vuorela

This value is the preedit-local coordinate of the event, which is
otherwise not reconstructable on the server from the pixel
coordinates.  MTF widgets report this value via the
preeditInjectionEvent, but this is the only way to get it for Qt
components.

Exposed as MInputMethodHost::preeditClickPos() in the server.

Signed-off-by: Andy Ross <andy.ross@windriver.com>

Full, original commit at 33f6c2d12e9c9638bf145735df07c679fb35e32b in maliit-framework