Fix failing Ut_MInputContext::testImInitiatedHideUnfocusesWidget
RevBy: TrustMe
Would fail on device, but not in scratchbox or host system.
This increases the wait delay to 1000ms.
As there is no better way to test this, and the results
look to be reliable, the test gets to live even if racy.

Full, original commit at f14198555f104738b652dffdb6573a4148c9568b in maliit-framework