Avoid use of QWidget::winId on win32
RevBy: TrustMe
We treat it as something that can be cast to uint / ulonglong,
which does not work on Windows. The code also does not make sense
there as the functionality it is used for is X11 specific.

Full, original commit at 3fee727f2e6021058c9fd958d85d505063861d7f in maliit-framework